requirements.txt
Flask>=2.2.0
flask_mysqldb>=1.0.1
mysql-connector-python>=8.0.0 